# Conflicts:
#	External/Plugins/ASCompletion/Completion/ASComplete.cs
#	External/Plugins/ASCompletion/Completion/ASGenerator.cs
#	External/Plugins/ProjectManager/Controls/TreeView/WatcherNode.cs
#	FlashDevelop/Managers/ScintillaManager.cs
#	PluginCore/ScintillaNet/Configuration/ConfigurationUtility.cs

+++ PluginCore/ScintillaNet/Configuration/ConfigurationUtility.cs
@@ -91,10 +91,13 @@ namespace ScintillaNet.Configuration
                         String tabContent = overriding.Replace("\n", "\n\t\t\t");
                         Int32 indexStart = original.IndexOfOrdinal(coloringStart);
                         Int32 indexEnd = original.IndexOfOrdinal(coloringEnd);
-                        String replaceTarget = original.Substring(indexStart, indexEnd - indexStart + coloringEnd.Length);
-                        String finalContent = original.Replace(replaceTarget, tabContent);
-                        File.WriteAllText(filename, finalContent);
-                        File.Delete(filename + ".override");
+                        if (indexStart > -1)
+                        {
+                            String replaceTarget = original.Substring(indexStart, indexEnd - indexStart + coloringEnd.Length);
+                            String finalContent = original.Replace(replaceTarget, tabContent);
+                            File.WriteAllText(filename, finalContent);
+                            File.Delete(filename + ".override");
+                        }
                     }
                     catch { /* NO ERRORS... */ }
                 }
